move all options to a separate module
